Improve default "overhead ratios".
The mapping from segment ordinals to global ordinals is often the bottleneck,
so it makes sense to make it configurable with a high overhead ratio. However,
the other mappings are never the bottleneck so we'd rather save memory on them.

Besides, FAST is a better default than FASTEST since the latter might be very
wasteful when only few bits per value are required.